mnist_backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
tensorflow-cpu>=2.13.0
Pillow>=10.0.0
python-multipart>=0.0.6
pydantic>=2.0.0
scikit-learn>=1.3.0
